DRTII-1741 Tidying & error fix
