+++ src/components/Shifts/EditShiftForm.tsx
@@ -1,131 +1,10 @@
 import React, {useState} from 'react';
-import {Box, Button, Grid, IconButton, MenuItem, Select, TextField, ThemeProvider, Typography} from '@mui/material';
+import {Box, Grid, IconButton, MenuItem, Select, TextField, Typography} from '@mui/material';
 import {intervalEndTimeOptions, intervalStartTimeOptions} from '../Util';
-import {ConfirmShiftForms} from "./ConfirmShiftForms";
-import {drtTheme} from "../../index";
 import CloseIcon from "@mui/icons-material/Close";
-import {getAirportNameByCode} from "../../aiports";
-import AddIcon from '@mui/icons-material/Add';
+import {ShiftForm} from "./AddShiftsForm";
 
-export interface ShiftForm {
-  id: number;
-  name: string;
-  startTime: string;
-  endTime: string;
-  defaultStaffNumber: number;
-}
-
-export interface ShiftsFormProps {
-  port: string;
-  terminal: string;
-  interval: number;
-  shiftForms: ShiftForm[];
-  confirmHandler: (shiftForms: ShiftForm[]) => void;
-}
-
-export const AddShiftForm = ({port, terminal, interval, shiftForms, confirmHandler}: ShiftsFormProps) => {
-  const [shifts, setShifts] = useState<ShiftForm[]>(Array.from(shiftForms).length > 0 ? shiftForms : [
-    {id: 1, name: '', startTime: '', endTime: '', defaultStaffNumber: 0}
-  ]);
-  const [showConfirm, setShowConfirm] = useState(false);
-  const [showErrors, setShowErrors] = useState(false);
-
-  const onContinue = () => {
-    if (shiftsHaveErrors(shifts)) {
-      setShowErrors(true);
-      return;
-    }
-    setShowConfirm(true)
-  };
-
-  const onCancel = () => {
-    setShowConfirm(false);
-  };
-
-  const shiftIsValid = (shift: ShiftForm) => {
-    return shift.name !== '' &&
-      shift.startTime !== '' &&
-      shift.endTime !== '' &&
-      shift.startTime < shift.endTime
-  }
-
-  const shiftsHaveErrors = (shifts: ShiftForm[]) => shifts.some(shift => {
-    console.log(`shiftIsValid: ${shiftIsValid(shift)} ${JSON.stringify(shift, null, 2)}`)
-    return !shiftIsValid(shift)
-  })
-
-  const handleAddShift = () => {
-    if (shiftsHaveErrors(shifts)) {
-      setShowErrors(true);
-      return;
-    }
-
-    setShifts([
-      ...shifts,
-      {id: shifts.length + 1, name: '', startTime: '', endTime: '', defaultStaffNumber: 0}
-    ]);
-    setShowErrors(false);
-  };
-
-  const handleRemoveShift = (id: number) => {
-    setShifts(shifts.filter(shift => shift.id !== id));
-  };
-
-  return (
-    <ThemeProvider theme={drtTheme}>
-      <Box>
-        {!showConfirm ? (
-          <Box sx={{p: 2, minWidth: '500px'}}>
-            <Typography sx={{fontSize: '20px'}}>Add staff to {port} {getAirportNameByCode(port)} {terminal}</Typography>
-            <Typography variant="h1" sx={{paddingBottom: '10px'}}>Step 1 of 2 - Create your shift pattern</Typography>
-            {Array.from(shifts).map(form => {
-              return <EditShiftForm key={form.id}
-                                    formState={form}
-                                    onUpdate={state => {
-                                      const updatedShifts = shifts.map(shift => {
-                                        if (shift.id === state.id) {
-                                          return {...shift, ...state};
-                                        }
-                                        return shift;
-                                      });
-                                      setShifts(updatedShifts)
-                                      setShowErrors(shiftsHaveErrors(updatedShifts))
-                                    }}
-                                    interval={interval}
-                                    removeShift={handleRemoveShift}
-                                    showSubmitErrors={showErrors}
-              />
-            })}
-            <Box>
-              <Button variant="outlined" color="primary" onClick={handleAddShift} sx={{gap: 0, paddingLeft: '0'}}>
-                <IconButton color="primary" sx={{padding: '0'}}>
-                  <AddIcon/>
-                </IconButton>
-                Add a shift
-              </Button>
-            </Box>
-            <Box sx={{"paddingTop": "10px"}}>
-              <Button variant="contained" color="primary" onClick={onContinue} data-cy="shift-continue-button">
-                Continue
-              </Button>
-              {showErrors && (
-                <Typography color="error" variant="body2">Please fix the errors before continuing</Typography>
-              )}
-            </Box>
-          </Box>) : (
-          <ConfirmShiftForms port={port}
-                             terminal={terminal}
-                             shifts={shifts}
-                             editShiftsHandler={onCancel}
-                             confirmHandler={confirmHandler}
-                             removeShiftHandler={handleRemoveShift}/>)
-        }
-      </Box>
-    </ThemeProvider>
-  )
-};
-
-const EditShiftForm = ({formState, onUpdate, interval, removeShift, showSubmitErrors}: {
+export const EditShiftForm = ({formState, onUpdate, interval, removeShift, showSubmitErrors}: {
   formState: ShiftForm,
   onUpdate: (state: ShiftForm) => void,
   interval: number,
@@ -146,7 +25,7 @@ const EditShiftForm = ({formState, onUpdate, interval, removeShift, showSubmitEr
     const updatedForm = {...formState, startTime: newStartTime};
     onUpdate(updatedForm);
     console.log(`formstate: ${JSON.stringify(updatedForm, null, 2)}`)
-    setTimeError(updatedForm.startTime !== '' && updatedForm.endTime !== '' && updatedForm.endTime <= updatedForm.startTime)
+    setTimeError(updatedForm.startTime !== '' && updatedForm.endTime !== '' && updatedForm.endTime == updatedForm.startTime)
     setStartTimeError(!startTimeOptions.includes(updatedForm.startTime))
   };
 
@@ -155,7 +34,7 @@ const EditShiftForm = ({formState, onUpdate, interval, removeShift, showSubmitEr
     const updatedForm = {...formState, endTime: newEndTime};
     onUpdate(updatedForm);
     console.log(`formstate: ${JSON.stringify(updatedForm, null, 2)}`)
-    setTimeError(updatedForm.startTime !== '' && updatedForm.endTime !== '' && updatedForm.endTime <= updatedForm.startTime)
+    setTimeError(updatedForm.startTime !== '' && updatedForm.endTime !== '' && updatedForm.endTime == updatedForm.startTime)
     setEndTimeError(!startTimeOptions.includes(updatedForm.endTime))
   };
 
@@ -174,9 +53,9 @@ const EditShiftForm = ({formState, onUpdate, interval, removeShift, showSubmitEr
       <Typography color="error" variant="body2">Please select end time</Typography>
     )}
     {timeError && (
-      <Typography color="error" variant="body2">Start time must be before the and end time</Typography>
+      <Typography color="error" variant="body2">Start time and end time cannot be the same</Typography>
     )}
-    {formState.endTime < formState.startTime && (
+    {formState.endTime && formState.startTime && formState.endTime < formState.startTime && (
       <Typography color="warning" variant="body2">This is a midnight shift spanning to the next day</Typography>
     )}
     <Grid container spacing={2}>
