+++ src/components/Shifts/index.ts
@@ -1,7 +1,7 @@
 export type {AddShiftBarProps} from './AddShiftBar'
 export {AddShiftBar} from "./AddShiftBar";
-export type {ShiftForm, ShiftsFormProps} from './AddShiftForm'
-export {AddShiftForm} from "./AddShiftForm";
+export type {ShiftForm, ShiftsFormProps} from './AddShiftsForm'
+export {AddShiftsForm} from "./AddShiftsForm";
 export type {ShiftDate, ShiftSummaryStaffing, ShiftSummary, StaffTableEntry, ShiftHotTableViewProps} from './ShiftHotTableView'
 export {ShiftHotTableView} from "./ShiftHotTableView";
 export type {ShiftSummaryProps} from './ShiftSummaryView'
